move Tile class into EightQueen.java
